requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
cramjam==2.5.0
cycler==0.11.0
Django==4.0.4
django-cors-headers==3.12.0
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
fastparquet==0.8.1
fonttools==4.33.3
fsspec==2022.5.0
greenlet==1.1.2
kiwisolver==1.4.2
matplotlib==3.5.2
mysqlclient==2.1.0
numpy==1.22.4
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
plotly==5.8.0
pyarrow==8.0.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
sqlparse==0.4.2
tenacity==8.0.1
tzdata==2022.1
